Moved invaliddataui to dom and updated index
