+++ client/dom/index.ts
@@ -11,6 +11,7 @@ export * from './fillbar'
 export * from './genesearch'
 export * from './GeneSetEdit/GeneSetEditUI'
 export { default as htmlLegend } from './html.legend'
+export { default as invaliddataui } from './invaliddataui.ts'
 export * from './LegendCircleReference.ts'
 export * from './maxLabelWidth'
 export * from './menu'
