+++ client/dom/invaliddataui.ts
